tools: use Jest for agent skill setup
