--- tools/agents/test/setup-skills.test.mjs
@@ -1,186 +1,0 @@
-import assert from 'node:assert/strict';
-import { execFile } from 'node:child_process';
-import {
-	mkdtemp,
-	mkdir,
-	readFile,
-	rm,
-	symlink,
-	writeFile,
-} from 'node:fs/promises';
-import os from 'node:os';
-import path from 'node:path';
-import test from 'node:test';
-import { fileURLToPath, pathToFileURL } from 'node:url';
-import { promisify } from 'node:util';
-
-import { setupSkills } from '../setup-skills.mjs';
-
-const execFileAsync = promisify( execFile );
-const setupScript = fileURLToPath(
-	new URL( '../setup-skills.mjs', import.meta.url )
-);
-const targetSkillPath = path.join( '.claude', 'skills', 'testing' );
-
-async function createRepository() {
-	const repositoryRoot = await mkdtemp(
-		path.join( os.tmpdir(), 'gutenberg-agent-skills-' )
-	);
-	const skillDirectory = path.join(
-		repositoryRoot,
-		'.agents/skills',
-		'testing'
-	);
-
-	await mkdir( skillDirectory, { recursive: true } );
-	await writeFile(
-		path.join( skillDirectory, 'SKILL.md' ),
-		'---\nname: testing\n'
-	);
-
-	return repositoryRoot;
-}
-
-test( 'links each canonical skill into the Claude directory', async ( t ) => {
-	const repositoryRoot = await createRepository();
-	t.after( () => rm( repositoryRoot, { recursive: true, force: true } ) );
-
-	const result = await setupSkills( { repositoryRoot } );
-
-	assert.deepEqual( result, {
-		conflicts: [],
-		copied: [],
-		linked: [ targetSkillPath ],
-	} );
-	assert.equal(
-		await readFile(
-			path.join( repositoryRoot, '.claude/skills/testing/SKILL.md' ),
-			'utf8'
-		),
-		'---\nname: testing\n'
-	);
-} );
-
-test( 'runs the setup command when executed directly', async ( t ) => {
-	const repositoryRoot = await createRepository();
-	t.after( () => rm( repositoryRoot, { recursive: true, force: true } ) );
-
-	const { stdout } = await execFileAsync( process.execPath, [ setupScript ], {
-		cwd: repositoryRoot,
-	} );
-
-	assert.equal( stdout, `Linked: ${ targetSkillPath }\n` );
-	assert.equal(
-		await readFile(
-			path.join( repositoryRoot, '.claude/skills/testing/SKILL.md' ),
-			'utf8'
-		),
-		'---\nname: testing\n'
-	);
-} );
-
-test( 'can be imported from an eval entrypoint', async () => {
-	await execFileAsync( process.execPath, [
-		'--input-type=module',
-		'--eval',
-		`await import( ${ JSON.stringify(
-			pathToFileURL( setupScript ).href
-		) } );`,
-	] );
-} );
-
-test( 'uses junctions on Windows', async ( t ) => {
-	const repositoryRoot = await createRepository();
-	t.after( () => rm( repositoryRoot, { recursive: true, force: true } ) );
-	const linkTypes = [];
-
-	await setupSkills( {
-		repositoryRoot,
-		platform: 'win32',
-		createLink: async ( source, target, type ) => {
-			linkTypes.push( type );
-			await symlink( source, target, type );
-		},
-	} );
-
-	assert.deepEqual( linkTypes, [ 'junction' ] );
-} );
-
-test( 'replaces links generated from the former skills directory', async ( t ) => {
-	const repositoryRoot = await createRepository();
-	t.after( () => rm( repositoryRoot, { recursive: true, force: true } ) );
-	const oldSkillDirectory = path.join( repositoryRoot, 'skills/testing' );
-	const target = path.join( repositoryRoot, '.claude/skills/testing' );
-
-	await mkdir( oldSkillDirectory, { recursive: true } );
-	await mkdir( path.dirname( target ), { recursive: true } );
-	await symlink(
-		oldSkillDirectory,
-		target,
-		process.platform === 'win32' ? 'junction' : 'dir'
-	);
-	await rm( oldSkillDirectory, { recursive: true, force: true } );
-	const result = await setupSkills( { repositoryRoot } );
-
-	assert.deepEqual( result.linked, [ targetSkillPath ] );
-	assert.equal(
-		await readFile( path.join( target, 'SKILL.md' ), 'utf8' ),
-		'---\nname: testing\n'
-	);
-} );
-
-test( 'copies when links are unavailable and refreshes only managed copies', async ( t ) => {
-	const repositoryRoot = await createRepository();
-	t.after( () => rm( repositoryRoot, { recursive: true, force: true } ) );
-	const unavailableLinks = async () => {
-		const error = new Error( 'Links are unavailable.' );
-		error.code = 'EPERM';
-		throw error;
-	};
-
-	const firstResult = await setupSkills( {
-		repositoryRoot,
-		createLink: unavailableLinks,
-	} );
-	assert.deepEqual( firstResult.copied, [ targetSkillPath ] );
-
-	await writeFile(
-		path.join( repositoryRoot, '.agents/skills/testing/SKILL.md' ),
-		'updated skill\n'
-	);
-	const refreshedResult = await setupSkills( {
-		repositoryRoot,
-		createLink: unavailableLinks,
-	} );
-
-	assert.deepEqual( refreshedResult.copied, [ targetSkillPath ] );
-	assert.equal(
-		await readFile(
-			path.join( repositoryRoot, '.claude/skills/testing/SKILL.md' ),
-			'utf8'
-		),
-		'updated skill\n'
-	);
-} );
-
-test( 'leaves an existing unmanaged skill untouched', async ( t ) => {
-	const repositoryRoot = await createRepository();
-	t.after( () => rm( repositoryRoot, { recursive: true, force: true } ) );
-	const unmanagedSkill = path.join(
-		repositoryRoot,
-		'.claude/skills/testing'
-	);
-
-	await mkdir( unmanagedSkill, { recursive: true } );
-	await writeFile(
-		path.join( unmanagedSkill, 'SKILL.md' ),
-		'personal skill\n'
-	);
-	const result = await setupSkills( { repositoryRoot } );
-
-	assert.deepEqual( result.conflicts, [ targetSkillPath ] );
-	assert.equal(
-		await readFile( path.join( unmanagedSkill, 'SKILL.md' ), 'utf8' ),
-		'personal skill\n'
-	);
-} );
